requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
requests==2.32.4
pandas==2.2.1
openpyxl==3.1.2
pillow==11.3.0
pytesseract==0.3.10
pdf2image==1.17.0
sendgrid==6.11.0
python-http-client==3.3.7
python-multipart==0.0.19